WiX: only setup SDKROOT for native platform
We should only setup `SDKROOT` for the native SDK.  This will be done on
x64 when installing the x64 SDK and on ARM64 when installing the ARM64
SDK.  If the user is cross-compiling, they should specify the SDKROOT by
passing `-sdk` or by setting up the SDK in the destination.json or by
manually setting the environment variable.

Use newer WiX features to detect the native machine type as
`%PROCESSOR_ARCHITECTURE` may not be sufficiently accurate.