WiX: move and rename toolchains on Windows
Migrate the default toolchain location from `%SystemDrive%` to
`%ProgramFiles%\Swift\Toolchains` before beginning the work to split up
the toolchain components.

Because the toolchain is now installed into a vendor indicating path
(`Swift`), it seems unnecessary to add the triple-compliant vendor
spelling into the toolchain identifier.  Dropping the `unknown-` prefix
leaves `Asserts-[version]`.  Rewrite this into a
[SemVer](https://semver.org) compliant spelling of `[version]+Asserts`.
We simply propagate the previous `0.0.0` version number as the spelling
of `development`.  This makes a main snapshot and local development
version identifier `0.0.0+Asserts`.

The preservation of the `Asserts` as a build modifier is to help enable
the parallel installation of a (unsupported) `NoAsserts` variant in case
a small performance gain is worth the risk of invalid code generation.

This installation naming matches the behaviour of Microsoft and .NET
Core.