WiX: wire up `ProductPlatform` to mirror `ProductArchitecture`
This wires up the product platform which is required to differentiate
the different SDKs being built. Because each SDK build references (and
builds) the shared wixlib, which was previously only parameterised on
the `ProductArchitecture`, we would hit build collisions when building
with sufficiently many threads and configurations. This adds another
axis of separation allowing safer parallel execution.

Incidentally, this work is also required to deal with the unified
installer build which cannot take all the requisite SDKROOT paths
in a single invocation.